Refactor compatibility dependencies
