+++ pytest_bdd/compatibility/pytest.py
@@ -5,7 +5,8 @@ from __future__ import annotations
 
 import sys
 from operator import ge
-from typing import TYPE_CHECKING
+from pathlib import Path
+from typing import TYPE_CHECKING, cast
 
 from _pytest.config import Config, PytestPluginManager
 from _pytest.config.argparsing import Parser
@@ -16,6 +17,7 @@ from _pytest.python import Metafunc
 from _pytest.reports import TestReport
 from _pytest.runner import CallInfo
 from _pytest.terminal import TerminalReporter
+from pytest import Module as PytestModule
 
 from pytest_bdd.packaging import compare_distribution_version
 
@@ -29,6 +31,8 @@ if compare_distribution_version("pytest", "7.0", ge):
     if TYPE_CHECKING:  # pragma: no cover
         from pytest import Testdir
 else:
+    import py
+
     if TYPE_CHECKING:  # pragma: no cover
         from _pytest.pytester import Testdir  # type: ignore[no-redef, attr-defined]
 
@@ -55,25 +59,84 @@ else:
 
 
 __all__ = [
-    "FixtureRequest",
     "Item",
-    "RunResult",
-    "Testdir",
-    "TypeAlias",
     "CallInfo",
     "call_fixture_func",
     "Config",
     "FixtureDef",
     "FixtureLookupError",
+    "FixtureRequest",
+    "get_config_root_path",
     "Mark",
     "MarkDecorator",
     "Metafunc",
+    "Module",
     "Parser",
     "PytestPluginManager",
     "PYTEST6",
     "PYTEST7",
+    "RunResult",
     "Session",
     "TerminalReporter",
+    "Testdir",
     "TestReport",
+    "TypeAlias",
     "wrap_session",
 ]
+
+
+class Module(PytestModule):
+    @classmethod
+    def build(cls, parent, file_path):
+        if hasattr(cls, "from_parent"):
+            collector = cls.from_parent(
+                parent, **(dict(path=Path(file_path)) if PYTEST7 else dict(fspath=py.path.local(file_path)))
+            )
+        else:
+            collector = cls(parent=parent, fspath=py.path.local(file_path))
+        return collector
+
+    def get_path(self):
+        return getattr(self, "path", Path(self.fspath))
+
+
+if PYTEST6:
+
+    def assert_outcomes(
+        result: RunResult,
+        passed: int = 0,
+        skipped: int = 0,
+        failed: int = 0,
+        errors: int = 0,
+        xpassed: int = 0,
+        xfailed: int = 0,
+    ) -> None:
+        """Compatibility function for result.assert_outcomes"""
+        result.assert_outcomes(
+            errors=errors, passed=passed, skipped=skipped, failed=failed, xpassed=xpassed, xfailed=xfailed
+        )
+
+else:
+
+    def assert_outcomes(
+        result: RunResult,
+        passed: int = 0,
+        skipped: int = 0,
+        failed: int = 0,
+        errors: int = 0,
+        xpassed: int = 0,
+        xfailed: int = 0,
+    ) -> None:
+        """Compatibility function for result.assert_outcomes"""
+        result.assert_outcomes(
+            error=errors,  # Pytest < 6 uses the singular form
+            passed=passed,
+            skipped=skipped,
+            failed=failed,
+            xpassed=xpassed,
+            xfailed=xfailed,
+        )
+
+
+def get_config_root_path(config: Config) -> Path:
+    return Path(getattr(cast(Config, config), "rootpath" if PYTEST61 else "rootdir"))
