+++ pytest_bdd/model/__init__.py
@@ -1,4 +1,5 @@
 from pytest_bdd.model.gherkin_document import Feature
+from pytest_bdd.model.messages import Type as StepType
 from pytest_bdd.model.scenario import Pickle, Step, Tag, UserStep
 
-__all__ = ["Feature", "Pickle", "Step", "Tag", "UserStep"]
+__all__ = ["Feature", "Pickle", "Step", "StepType", "Tag", "UserStep"]
