+++ setup.cfg
@@ -30,6 +30,7 @@ python_requires = >=3.7
 install_requires =
     aiohttp
     attrs
+    certifi
     cucumber-expressions
     gherkin-official>=24
     importlib-metadata;python_version<"3.10.0"
@@ -51,9 +52,9 @@ install_requires =
 tests_require = tox
 packages =
     pytest_bdd
+    pytest_bdd.compatibility
     pytest_bdd.model
     pytest_bdd.struct_bdd
-    pytest_bdd.typing
 include_package_data = True
 
 [options.extras_require]
